Refactor rest api endpoints
