+++ pages/api/elements/index.ts
@@ -1,0 +1,7 @@
+import { NextApiRequest, NextApiResponse } from 'next';
+import { createScenarioGroup, getScenarioGroups, createElement } from '../../../lib/rest-methods';
+
+export default async (req: NextApiRequest, res: NextApiResponse) => {
+  if (req.method === 'POST') return await createElement(req, res);
+  else return res.status(405).end();
+};
