+++ tsconfig.json
@@ -19,6 +19,6 @@
     "jsx": "preserve",
     "types": ["node"]
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "pages/api/element/delete.ts'"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "pages/api/elements/delete.ts'"],
   "exclude": ["node_modules"]
 }
